Roll Mean

Apply a moving-window mean function to a numeric vector.

roll_mean(x, n = 5L, weights = NULL, by = 1L, align = 0L)

Arguments

x
A numeric vector.

n
An integer window length.

weights
A numeric vector of size n specifying each window index weight. If NULL, the unit weight is used.

by
An integer to shift the window by.

align
A signed integer representing the windows alignment. -1(left)|0(center)|1(right).

Value

numeric vector of length(x)

Details

Each window of n-length is applied weight and then slid/shifted/rolled by a positive integer amount about the window's align-ment index.

Examples

# load airquality
data("airquality")

# calculate moving average of last 6 measurements
roll_mean(airquality$Temp, n = 6, align = -1)

#>   [1] 66.16667 65.83333 63.66667 61.50000 62.66667 65.66667 66.16667 66.33333
#>   [9] 67.83333 67.33333 66.50000 65.16667 63.16667 63.50000 62.50000 62.66667
#>  [17] 64.16667 63.33333 64.00000 62.16667 61.50000 61.16667 60.16667 63.50000
#>  [25] 66.50000 69.66667 73.00000 75.83333 75.83333 76.33333 77.33333 77.83333
#>  [33] 78.50000 80.66667 84.50000 85.00000 86.33333 88.50000 88.50000 87.33333
#>  [41] 85.50000 83.83333 80.33333 75.83333 74.33333 73.66667 73.33333 73.16667
#>  [49] 73.83333 75.66667 76.00000 76.33333 75.66667 76.33333 76.50000 77.66667
#>  [57] 79.16667 80.33333 81.66667 82.33333 83.33333 83.33333 84.00000 85.16667
#>  [65] 87.00000 87.83333 87.66667 86.00000 84.83333 84.66667 82.66667 81.33333
#>  [73] 81.33333 83.16667 84.16667 83.16667 82.16667 82.16667 82.16667 82.50000
#>  [81] 82.16667 81.66667 83.66667 84.83333 85.50000 85.00000 84.33333 84.16667
#>  [89] 83.33333 82.33333 82.33333 82.66667 83.66667 85.00000 86.50000 87.83333
#>  [97] 88.83333 89.00000 88.83333 87.66667 86.00000 84.16667 81.66667 80.50000
#> [105] 78.83333 78.16667 77.83333 77.50000 76.66667 76.00000 76.50000 77.00000
#> [113] 78.33333 80.16667 84.33333 87.50000 90.33333 92.50000 93.33333 94.00000
#> [121] 93.33333 93.16667 91.66667 90.00000 88.16667 85.83333 82.83333 79.50000
#> [129] 78.50000 77.16667 76.66667 75.50000 74.83333 75.66667 73.33333 73.33333
#> [137] 71.83333 73.66667 72.50000 71.33333 73.66667 72.50000 71.66667 69.66667
#> [145] 71.83333 72.50000 71.66667 71.50000       NA       NA       NA       NA
#> [153]       NA