Roll Product

Apply a moving-window product function to a numeric vector.

roll_prod(x, n = 5L, weights = NULL, by = 1L, align = 0L)

Arguments

x
A numeric vector.

n
An integer window length.

weights
A numeric vector of size n specifying each window index weight. If NULL, the unit weight is used.

by
An integer to shift the window by.

align
A signed integer representing the windows alignment. -1(left)|0(center)|1(right).

Value

numeric vector of length(x)

Details

Each window of n-length is applied weight and then slid/shifted/rolled by a positive integer amount about the window's align-ment index.

Examples

# load airquality
data("airquality")

# calculate moving product of 12 measurements
roll_prod(airquality$Temp, n = 12)

#>   [1]           NA           NA           NA           NA           NA
#>   [6]           NA 6.741985e+21 6.641358e+21 6.272394e+21 4.916201e+21
#>  [11] 5.074788e+21 5.981000e+21 5.165409e+21 5.403812e+21 5.678583e+21
#>  [16] 5.492400e+21 5.810800e+21 4.789983e+21 4.234623e+21 3.657174e+21
#>  [21] 3.119355e+21 3.065573e+21 3.209271e+21 3.938651e+21 5.458832e+21
#>  [26] 6.101048e+21 7.675512e+21 9.626913e+21 8.835660e+21 1.216714e+22
#>  [31] 1.695421e+22 2.349794e+22 3.322122e+22 5.070608e+22 6.811264e+22
#>  [36] 7.315803e+22 8.612274e+22 1.042538e+23 1.096002e+23 1.184867e+23
#>  [41] 1.397082e+23 1.280658e+23 1.084793e+23 8.925512e+22 7.945883e+22
#>  [46] 6.941231e+22 5.938609e+22 5.187750e+22 4.239452e+22 3.502156e+22
#>  [51] 3.203191e+22 3.123112e+22 2.885913e+22 2.998351e+22 3.206570e+22
#>  [56] 4.094543e+22 4.711529e+22 5.269474e+22 5.543213e+22 6.126709e+22
#>  [61] 6.691011e+22 7.307288e+22 8.573885e+22 1.011279e+23 1.274488e+23
#>  [66] 1.417868e+23 1.509938e+23 1.328017e+23 1.280588e+23 1.370983e+23
#>  [71] 1.354057e+23 1.305698e+23 1.289966e+23 1.305508e+23 1.290673e+23
#>  [76] 1.192470e+23 9.591603e+22 8.729436e+22 8.729436e+22 1.028399e+23
#>  [81] 1.079184e+23 9.724519e+22 1.045386e+23 1.135728e+23 1.191129e+23
#>  [86] 1.176949e+23 1.095780e+23 1.044214e+23 1.142991e+23 1.157102e+23
#>  [91] 1.213546e+23 1.199435e+23 1.227657e+23 1.332457e+23 1.394432e+23
#>  [96] 1.426123e+23 1.525620e+23 1.580763e+23 1.678341e+23 1.699062e+23
#> [101] 1.678085e+23 1.616692e+23 1.447503e+23 1.345327e+23 1.175228e+23
#> [106] 1.029975e+23 8.926450e+22 7.637074e+22 5.976841e+22 5.212361e+22
#> [111] 4.788099e+22 4.729708e+22 5.084436e+22 5.663675e+22 7.134760e+22
#> [116] 8.489461e+22 1.072353e+23 1.292323e+23 1.507710e+23 1.801420e+23
#> [121] 2.326834e+23 2.885274e+23 3.177453e+23 3.295137e+23 3.065244e+23
#> [126] 2.716920e+23 2.100712e+23 1.631404e+23 1.376497e+23 1.112912e+23
#> [131] 9.416951e+22 7.267430e+22 5.548253e+22 4.653373e+22 3.583632e+22
#> [136] 3.242334e+22 2.755984e+22 2.897316e+22 2.472377e+22 2.404640e+22
#> [141] 2.404640e+22 2.183160e+22 1.786222e+22 1.761064e+22 1.909886e+22
#> [146] 1.836429e+22 2.083114e+22           NA           NA           NA
#> [151]           NA           NA           NA