Roll Standard Deviation

Apply a moving-window standard deviation function to a numeric vector.

roll_sd(x, n = 5L, weights = NULL, by = 1L, align = 0L)

Arguments

x
A numeric vector.

n
An integer window length.

weights
A numeric vector of size n specifying each window index weight. If NULL, the unit weight is used.

by
An integer to shift the window by.

align
A signed integer representing the windows alignment. -1(left)|0(center)|1(right).

Value

numeric vector of length(x)

Details

Each window of n-length is applied weight and then slid/shifted/rolled by a positive integer amount about the window's align-ment index.

Examples

# load airquality
data("airquality")

# calculate moving standard deviation of adjacent measurements
roll_mean(airquality$Temp, n = 3)

#>   [1]       NA 71.00000 69.33333 64.00000 61.33333 62.33333 63.33333 61.66667
#>   [9] 63.00000 68.00000 70.66667 69.66667 67.66667 64.00000 63.33333 62.66667
#>  [17] 62.33333 63.66667 62.33333 63.00000 64.66667 64.33333 65.00000 59.66667
#>  [25] 58.66667 57.33333 60.66667 68.33333 75.66667 78.66667 77.66667 76.00000
#>  [33] 73.00000 75.00000 78.66667 82.66667 82.00000 82.66667 86.33333 88.00000
#>  [41] 90.00000 90.66667 89.00000 84.66667 80.33333 78.66667 76.00000 71.33333
#>  [49] 70.00000 71.33333 75.33333 76.33333 76.33333 76.00000 75.66667 76.33333
#>  [57] 75.33333 77.00000 76.66667 80.00000 81.33333 84.00000 83.33333 83.33333
#>  [65] 82.66667 83.33333 84.66667 87.66667 90.66667 91.00000 87.66667 81.33333
#>  [73] 78.66667 81.66667 84.00000 84.00000 81.00000 82.33333 84.33333 85.33333
#>  [81] 82.00000 80.00000 79.00000 83.00000 84.33333 84.33333 84.33333 85.33333
#>  [89] 86.66667 85.66667 83.33333 81.66667 81.00000 81.33333 83.00000 84.33333
#>  [97] 86.00000 87.00000 88.66667 89.66667 90.66667 89.33333 88.00000 84.66667
#> [105] 82.66667 80.33333 78.66667 78.33333 77.33333 77.66667 77.33333 77.66667
#> [113] 75.66667 74.66667 75.33333 78.33333 82.00000 85.00000 90.33333 93.00000
#> [121] 95.66667 94.66667 93.66667 92.33333 92.00000 92.66667 91.00000 88.00000
#> [129] 83.66667 80.66667 77.66667 75.33333 76.33333 76.66667 78.00000 74.66667
#> [137] 73.00000 73.33333 72.00000 73.66667 70.33333 75.33333 71.33333 72.33333
#> [145] 72.00000 73.66667 71.00000 67.33333 70.00000 74.00000 76.00000 73.00000
#> [153]       NA