Roll Sum

Apply a moving-window sum to a numeric vector.

roll_sum(x, n = 5L, weights = NULL, by = 1L, align = 0L)

Arguments

x
A numeric vector.

n
An integer window length.

weights
A numeric vector of size n specifying each window index weight. If NULL, the unit weight is used.

by
An integer to shift the window by.

align
A signed integer representing the windows alignment. -1(left)|0(center)|1(right).

Value

numeric vector of length(x)

Details

Each window of n-length is applied weight and then slid/shifted/rolled by a positive integer amount about the window's align-ment index.

Examples

# load airquality
data("airquality")

# calculate moving sum of last 3 measurements
roll_sum(airquality$Temp, n = 3, align = -1)

#>   [1] 213 208 192 184 187 190 185 189 204 212 209 203 192 190 188 187 191 187
#>  [19] 189 194 193 195 179 176 172 182 205 227 236 233 228 219 225 236 248 246
#>  [37] 248 259 264 270 272 267 254 241 236 228 214 210 214 226 229 229 228 227
#>  [55] 229 226 231 230 240 244 252 250 250 248 250 254 263 272 273 263 244 236
#>  [73] 245 252 252 243 247 253 256 246 240 237 249 253 253 253 256 260 257 250
#>  [91] 245 243 244 249 253 258 261 266 269 272 268 264 254 248 241 236 235 232
#> [109] 233 232 233 227 224 226 235 246 255 271 279 287 284 281 277 276 278 273
#> [127] 264 251 242 233 226 229 230 234 224 219 220 216 221 211 226 214 217 216
#> [145] 221 213 202 210 222 228 219  NA  NA